Move file to models
